feat(struct): changed struct of folders
The folder structure has been
changed with the aim of making it
simpler and more intuitive way
of allocating code thus modularizing
the code. The base files for
building the interface were also
the requirements were updated.
