+++ src/database/tables_db/subject.py
@@ -7,7 +7,11 @@ class Subject:
 
     def create_subject(self):
         query = """
-    MERGE (sub:Subject {subject_id: $subject_id, name: $name, teacher_id: $teacher_id})
-    """
-        parameters = {"subject_id": self.subject_id, "name": self.name, "teacher_id": self.teacher_id}
+                MERGE (sub:Subject {subject_id: $subject_id, name: $name, teacher_id: $teacher_id})
+                """
+        parameters = {
+            "subject_id": self.subject_id,
+            "name": self.name,
+            "teacher_id": self.teacher_id,
+        }
         self.db.execute_query(query, parameters)
