+++ src/database/tables_db/teacher.py
@@ -6,16 +6,16 @@ class Teacher:
 
     def create_teacher(self):
         query = """
-    MERGE (t:Teacher {teacher_id: $teacher_id, name: $name})
-    """
+                MERGE (t:Teacher {teacher_id: $teacher_id, name: $name})
+                """
         parameters = {"teacher_id": self.teacher_id, "name": self.name}
         self.db.execute_query(query, parameters)
 
     def add_subject(self, subject):
         query = """
-    MATCH (t:Teacher {teacher_id: $teacher_id})
-    MATCH (sub:Subject {subject_id: $subject_id})
-    MERGE (t)-[:TEACHES]->(sub)
-    """
+                MATCH (t:Teacher {teacher_id: $teacher_id})
+                MATCH (sub:Subject {subject_id: $subject_id})
+                MERGE (t)-[:TEACHES]->(sub)
+                """
         parameters = {"teacher_id": self.teacher_id, "subject_id": subject.subject_id}
         self.db.execute_query(query, parameters)
